backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.2",
    "groq>=0.31.0",
    "langgraph>=1.1.3",
    "pillow>=12.1.1",
    "pydantic>=2.11.0",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.42.0",
]
